Fix for: https://github.com/aloisdeniel/flutter_device_preview/issues/118
Exception thrown says framework wants didChangeDependencies, so I just tried giving it that instead of initState. Also the initState was no longer needed. In my use and test cases it removes the thrown exception and everything still seems to work. But I would of course still recommend taking a closer a look at the issue and run it through your tests.